feat(bo-htx): Add functions to encode and decode frames with length prefix, update documentation and tests in htx-lab to expose these functions
