+++ crates/bo-htx/src/lib.rs
@@ -1,5 +1,6 @@
-/// bo-htx: SANS-IO HTX tunnel primitives
+/// bo-htx: SANS-IO HTX tunnel primitives (canonical library)
 
+/// Encode a simple tunnel header (magic + version + flags + length)
 pub fn encode_tunnel_header(version: u8, flags: u8, payload_len: u16) -> [u8;8] {
     let mut h = [0u8;8];
     // magic 4 bytes
@@ -17,6 +18,24 @@ pub fn decode_tunnel_length(header: &[u8]) -> Option<u16> {
     Some(len)
 }
 
+/// Encode a simple length-prefixed frame
+pub fn encode_frame(payload: &[u8]) -> Vec<u8> {
+    let mut out = Vec::with_capacity(2 + payload.len());
+    let len = payload.len() as u16;
+    out.push((len >> 8) as u8);
+    out.push((len & 0xff) as u8);
+    out.extend_from_slice(payload);
+    out
+}
+
+/// Decode a single length-prefixed frame if present
+pub fn decode_frame(buf: &[u8]) -> Option<Vec<u8>> {
+    if buf.len() < 2 { return None; }
+    let len = ((buf[0] as usize) << 8) | (buf[1] as usize);
+    if buf.len() < 2 + len { return None; }
+    Some(buf[2..2+len].to_vec())
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -25,6 +44,14 @@ mod tests {
     fn encode_decode_header() {
         let h = encode_tunnel_header(1, 0, 5);
         assert_eq!(decode_tunnel_length(&h), Some(5));
+    }
+
+    #[test]
+    fn smoke_tunnel_frame() {
+        let payload = b"hello";
+        let enc = encode_frame(payload);
+        let dec = decode_frame(&enc).unwrap();
+        assert_eq!(dec, payload);
     }
 }
