+++ crates/htx-lab/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0"
 bytes = "1.4"
+bo-htx = { path = "../bo-htx" }
 
 [dev-dependencies]
 # Use cargo-fuzz as an external tool; do not list as optional dev-dep here.
