+++ crates/htx-lab/src/tunnel_mock.rs
@@ -1,19 +1,11 @@
-/// tunnel_mock: encapsulación / decapsulación simple (length-prefixed)
+use bo_htx::{encode_frame, decode_frame};
 
-pub fn encode_frame(payload: &[u8]) -> Vec<u8> {
-    let mut out = Vec::with_capacity(2 + payload.len());
-    let len = payload.len() as u16;
-    out.push((len >> 8) as u8);
-    out.push((len & 0xff) as u8);
-    out.extend_from_slice(payload);
-    out
+pub fn encode_frame_example(payload: &[u8]) -> Vec<u8> {
+    encode_frame(payload)
 }
 
-pub fn decode_frame(mut buf: &[u8]) -> Option<Vec<u8>> {
-    if buf.len() < 2 { return None; }
-    let len = ((buf[0] as usize) << 8) | (buf[1] as usize);
-    if buf.len() < 2 + len { return None; }
-    Some(buf[2..2+len].to_vec())
+pub fn decode_frame_example(buf: &[u8]) -> Option<Vec<u8>> {
+    decode_frame(buf)
 }
 
 #[cfg(test)]
@@ -23,8 +15,8 @@ mod tests {
     #[test]
     fn smoke_tunnel() {
         let p = b"hello";
-        let enc = encode_frame(p);
-        let dec = decode_frame(&enc).unwrap();
+        let enc = encode_frame_example(p);
+        let dec = decode_frame_example(&enc).unwrap();
         assert_eq!(dec, p);
     }
 }
